the Button.pde is updated
The 'Console' code also doesn't work properly in the KTByte Coder.
Turned out that the problem is with the KTByte Coder's own console (the one located at the bottom of the screen). Particularly, with the 'println()' method.
To check this, try to put any 'println()' statement inside the 'void onConsoleInput()' and you will not see any output in the KTByte console (in the bottom pane of the KTByte Coder). While the other code related to the callback itself will work.
For example, let's add some 'println()' statements:
```java
  console.setConsoleInputEvent(new ConsoleInputEvent() {

    void onConsoleInput(String variable, String value) {
      if (variable.equals("number")) {
        // using custom getInteger() method since using Processing's "int(...)" fails when Java code is compiling
        if (getInteger(value) < number) {
          console.write("Too small. Try again!");
          console.readInput("number");
          println("Too small. Try again!");
        } else if (getInteger(value) > number) {
          console.write("Too big. Try again!");
          console.readInput("number");
          println("Too big. Try again!");
        } else {
          console.write("Exactly!");
          println("Exactly!");
        }
      }
    }
  }
  );
```

But the most interesting is that the method of the anonymous class will work! As an example - we can change the color of the background each time the button is pressed (See the updated code inside the Button.pde).