Initial commit.
Added 'tools.jar' into the 'tools' folder and to the build path. Reason: simplify development for the person who will import this folder as a project into Eclipse.
Added 'core.jar' into the 'processing' folder and to the build path. Reason: simplify development for the person who will import this folder as a project into Eclipse.