Simple implementation of 'duplicated' mouse event processing
Previously, there was a software design issue in mouse events (such as dragging and pressing) processing for components that has childs attached to them (for example, Buttons and Sliders inside the Pane or Window). The mouse event processed for both, parent and childs. That cause the situations when both, parent and child, was pressed/dragged at the same time. This is fixed now. But only for the case of 'parent-child'. If some of the parent's chids is pressed, then the parent controller ignores the mouse event. That is, the child controller has higher priority in mouse event processing. IMPORTANT: For the case of 'child-child' overlapping inside the same parent, or the case 'controller-controller' (both without childs) overlapping, this not work and should be implemented later.