Fix how GUI controllers process mouse presses
The problem was the following: Some of the components were not 'ready'
to process next 'mousePressed' event if the mouse pointer hasn't
changed its position since the previous 'mousePressed' event. It was
such that the user had to move the mouse pointer to be able to press
the button (controller) again.