Constrain ScrollableTextArea scrolling
  Previously, when we scrolled the text with the ScrollBar, it was
possible to scroll such that the last line could appeared at the _top_
of the visible area. At the same time, using the MouseWheel, the
scrolling process resulted in a 'constrained' scrolling - i.e. the last
line is always a last line - at the _bottom_ of the visible area.
  This commit makes them (scrolling using MouseWheel and scroling using
ScrollBar) work the same - the last line of the text is alwasy at the
_bottom_.