Fix the last visible line in ScrollableTextArea
Previously, when we scrolled the text inside a visible area of the
ScrollableTextArea, we had to deal with incorrectly calculated number
of the last line to be showed inside the ScrollableTextArea. The
calculated value was non-constant (!). This prevents to correctly map
the 'currentLinePosition' to ScrollBar handle position and also caused
the visibility of an additional empty line at the end of the text.

The method which caused the error and which was fixed is
 ScrollableTextArea.incrementLine().