Synchronize Console's TextArea with Slider handle
This commit creates a one-way sync between the ScrollableTextArea's
text 'startLinePosition' and the ScrollBar's Slider handle position.
When the user scrolls the text inside the TextArea with a mouse
wheel, the Slider's handle position is moving appropriately.
The event adapters of the ScrollBar (subscribed to the changes of the
handle position with the use of an onValueChanged() method) are not
notified, including the ScrollableTextArea itself. This way, the two
events generated by these two GUI components do not interfere.
This is achieved by setting the 'isActive' variable of the ScrollBar
to false at this particular moment of time. That way, the handle of the
ScrollBar is changing and is being redrawn at a new position, but the
corresponding event is not sent to any subscribed adapter.